Implement HGPanel baseclass with warning method
