+++ user_interface/main_panel/pose_ui.py
@@ -1,7 +1,7 @@
 import addon_utils
 import bpy
 
-from .main_panel_baseclass import MainPanelPart, subpanel_draw
+from ..ui_baseclasses import MainPanelPart, subpanel_draw
 
 
 class HG_PT_POSE(MainPanelPart, bpy.types.Panel):
