+++ user_interface/main_panel/skin_ui.py
@@ -3,7 +3,7 @@ from sys import platform
 import bpy
 
 from ..panel_functions import draw_sub_spoiler
-from .main_panel_baseclass import MainPanelPart, subpanel_draw
+from ..ui_baseclasses import MainPanelPart, subpanel_draw
 
 
 class HG_PT_SKIN(MainPanelPart, bpy.types.Panel):
