+++ user_interface/panel_functions.py
@@ -73,7 +73,7 @@ def get_flow(sett, layout, animation=False) -> bpy.types.UILayout:
     return flow
 
 
-def draw_spoiler_box(self, ui_name) -> "tuple[bool, bpy.types.UILayout]":
+def draw_spoiler_box(self, layout, ui_name) -> "tuple[bool, bpy.types.UILayout]":
     """Draws the spoiler box of the main sections (i.e. body, hair, face)
 
     Args:
@@ -100,8 +100,6 @@ def draw_spoiler_box(self, ui_name) -> "tuple[bool, bpy.types.UILayout]":
         "simulation": "NETWORK_DRIVE",
         "compression": "FOLDER_REDIRECT",
     }
-
-    layout = self.col
     box = layout.box()
 
     row = box.row(align=True)
